// app/types/pocketbase-types.ts
export type IsoDateString = string
// ...
export enum JobsRemoteOptions {
	"💻 Remote" = "💻 Remote",
}
// ...
export enum JobsPostTypeOptions {
	"featured" = "featured",
	"aggregated" = "aggregated",
}
export type JobsRecord = {
	updated_ats: IsoDateString
	logo?: string
	title?: string
	company?: string
	country?: JobsCountryOptions[]
	remote?: JobsRemoteOptions
	tags?: JobsTagsOptions[]
	location?: string
	min_salary?: string
	max_salary?: string
	apply_link?: string
	id_ats?: string
	commitment?: string
	region?: string
	company_url?: string
	post_type?: JobsPostTypeOptions
	paid?: boolean
}